refactor: 404 페이지 관련 폴더 및 파일명 수정
