+++ src/pages/404/404.tsx
@@ -1,7 +1,8 @@
-import { MainContainer } from './error.styles';
+import { MainContainer } from './404.styles';
 import BottomButton from '../../components/bottomButton/BottomButton';
 import { useNavigate } from 'react-router-dom';
 import { ROUTES } from '../../constants/ROUTES';
+
 const ErrorPage = () => {
   const navigate = useNavigate();
   return (
